Fold the expiring cache into LatestRepository instead of adding a sibling
The style gate rejected a 29th file in packages/shared/cache, and chasing that
found the better answer: LatestRepository already is the synchronous push-side
keyed cache over LatestValue, implementing PushKeyedValues with accept, read,
take, deleteExpired, size and eviction options. ExpiringRepository duplicated
almost all of it and differed only in the three things the #240 analysis said
were missing.

Those three move onto LatestRepository directly: acceptAt with a per-entry
absolute deadline, readAt/expiredAt/deleteExpiredAt taking the deciding
instant, and evictExpiredWhenAllowed running the scan through RateLimiter on
the caller's clock. readOrAcceptAt covers the memoise case. The existing
wall-clock and setInterval paths are untouched, so current callers see no
change; the new surface is for callers that own a clock and have no lifecycle
hook to dispose a timer.

This removes the new file, so the directory stays at 28 and the ownership
finding is no longer worsened. It also leaves one keyed push repository rather
than two overlapping ones.
